// src/app/enums/PokemonTypeColor.ts
export enum PokemonTypeColor {
  Normal = '#AAAA99',
  Fire = '#FF4422',
  Water = '#3399FF',
  Electric = '#FFCC33',
  Grass = '#77CC55',
  Ice = '#66CCFF',
  Fighting = '#BB5544',
  Poison = '#AA5599',
  Ground = '#DDBB55',
  Flying = '#8899FF',
  Psychic = '#FB5497',
  Bug = '#AABB22',
  Rock = '#BBAA66',
  Ghost = '#6666BB',
  Dragon = '#7766EE',
  Dark = '#775544',
  Steel = '#AAAABB',
  Fairy = '#EE99EE',
}